+++ src/financial/index.ts
@@ -4,3 +4,4 @@ export * from "./investment";
 export * from "./rate";
 export * from "./simple";
 export * from "./leverage";
+export * from "./depreciation";